backend/requirements.txt
fastapi<0.100.0
uvicorn[standard]==0.23.1
openai==0.27.8
tiktoken==0.4.0
python-dotenv==1.0.0

langchain 
pinecone-client==2.2.2
google-cloud-firestore
lark
google-cloud-secret-manager